pyproject.toml
[tool.poetry]
name = "vidly-api"
version = "0.1.0"
description = ""
authors = ["Damilola Osunsanwo <oluwadamilola013@gmail.com>"]
readme = "README.md"
packages = [{include = "vidly_api"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
mysqlclient = "^2.2.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
